py_pkg_requirements.txt
'''
Created on August 17, 2020
@author: hmo (hyunho.mo@unitn.it)
python package requirements for pdm_dashboard
'''

pandas==1.1.0
numpy==1.19.1
scikit-learn==0.23.1
tqdm==4.48.2
tensorflow-gpu==2.1.0
deap==1.3.1
matplotlib==3.3.1
